refactor: modularize Token and list()
