--- lib/helpers.js
@@ -1,3 +1,0 @@
-export function unescape(identifier) {
-  return identifier.startsWith('_') ? identifier.slice(1) : identifier;
-}
