+++ lib/tokeniser.js
@@ -92,7 +92,7 @@ const punctuations = [
 /**
  * @param {string} str
  */
-export function tokenise(str) {
+function tokenise(str) {
   const tokens = [];
   let lastCharIndex = 0;
   let trivia = "";
